# be/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Hospital Management API",
        "title": "Hospital Management API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User's email and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiResponseSuccessStruct"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "Logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "User's refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiResponseSuccessStruct"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "description": "Refresh Access Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Access Token",
                "parameters": [
                    {
                        "description": "User's refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiResponseSuccessStruct"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ApiResponseSuccessStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ApiResponseSuccessStruct": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "dto.DoctorInfoRequest": {
            "type": "object",
            "required": [
                "medical_license_number",
                "specialization"
            ],
            "properties": {
                "medical_license_number": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LogoutRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "dto.ManagerInfoRequest": {
            "type": "object",
            "required": [
                "department"
            ],
            "properties": {
                "department": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.NurseInfoRequest": {
            "type": "object",
            "required": [
                "nursing_license_number"
            ],
            "properties": {
                "nursing_license_number": {
                    "type": "string"
                }
            }
        },
        "dto.PatientInfoRequest": {
            "type": "object",
            "properties": {
                "allergies": {
                    "type": "string"
                },
                "blood_type": {
                    "type": "string"
                },
                "chronic_conditions": {
                    "type": "string"
                },
                "insurance_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "address",
                "citizen_id",
                "date_of_birth",
                "email",
                "gender",
                "name",
                "password",
                "phone_number",
                "role"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "citizen_id": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "manager_info": {
                    "$ref": "#/definitions/dto.ManagerInfoRequest"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "patient_info": {
                    "$ref": "#/definitions/dto.PatientInfoRequest"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "staff_info": {
                    "$ref": "#/definitions/dto.StaffInfoRequest"
                }
            }
        },
        "dto.StaffInfoRequest": {
            "type": "object",
            "required": [
                "department",
                "role"
            ],
            "properties": {
                "department": {
                    "type": "string"
                },
                "doctor_info": {
                    "$ref": "#/definitions/dto.DoctorInfoRequest"
                },
                "nurse_info": {
                    "$ref": "#/definitions/dto.NurseInfoRequest"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}